backend/requirements.txt
fastapi[standard]
python-dotenv
python-jose[cryptography]
httpx
psycopg[binary]
psycopg[pool]

sqlalchemy
pydantic
python-dotenv
vercel
python-multipart
